+++ src/firebase/firebase.ts
@@ -15,5 +15,4 @@ const app = initializeApp({
 export const auth = getAuth(app);
 export const database = getFirestore(app);
 
-
 export default app;
